add fields and constructors
